➕ Add slots and the logic. 🔧 Fix the "can place item on grid" condition, to check if is not empty through "null" and not booleans. 🔧 Tweak on inventory logic.

+++ src/lib/models/InventoryIndexes.ts
@@ -21,7 +21,7 @@ const InventoryIndexSchema: JTDSchemaType<InventoryIndex> = {
 const InventoryBlockIndexSchema: JTDSchemaType<InventoryBlockIndexes> = {
     properties: {
         ...InventoryIndexSchema.properties,
-        block: { type: 'uint8' }
+        block: { type: 'int16' }
     }
 }
